// src/lib/enums/product.enum.ts
export enum ProductSize {
  SMALL = "SMALL",
  NORMAL = "NORMAL",
  LARGE = "LARGE",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum ProductCollection {
  FLOWERS = "FLOWERS",
  PLANTS = "PLANTS",
  ACCESSORIES = "ACCESSORIES",
  SEEDS = "SEEDS",
  OTHER = "OTHER",
}

export enum ProductVolume {
  HALF = 0.5,
  ONE = 1,
  TWO = 2,
  FIVE = 5,
}